[DASH-2365] Separate code into more components
