+++ src/apps/DataEndpoints/DataEndpointsPreview.js
@@ -6,13 +6,15 @@ import { withRouter } from 'react-router';
 
 import { DialogsMixin } from '../../mixins';
 import Constants from '../../constants/Constants';
+import DataObjectsTableInitialColumns from '../DataObjects/DataObjectsTableInitialColumns';
 
 import Actions from './DataEndpointsPreviewActions';
 import Store from './DataEndpointsPreviewStore';
 
 import { IconButton } from 'material-ui';
-import { Container, Dialog, InnerToolbar, Loading, DataObjectsTable } from '../../common';
+import { Container, Dialog, InnerToolbar, Loading } from '../../common';
 import ReadOnlyTooltip from '../DataObjects/ReadOnlyTooltip';
+import DataObjectsTable from '../DataObjects/DataObjectsTable';
 
 const DataEndpointsPreview = React.createClass({
   mixins: [
@@ -79,12 +81,6 @@ const DataEndpointsPreview = React.createClass({
     });
   },
 
-  handleMoreRows() {
-    const { nextLink } = this.state;
-
-    Actions.fetchNextDataPage(nextLink);
-  },
-
   handleDelete() {
     const { classObject, items, selectedRows } = this.state;
     const selectedDataObectsIds = _.map(selectedRows, (rowId) => items[rowId].id);
@@ -111,9 +107,9 @@ const DataEndpointsPreview = React.createClass({
   },
 
   render() {
-    const { isLoading, classObject, selectedRows, currentPage, hasNextPage, items } = this.state;
-    const { dataEndpointName } = this.props.params;
     const styles = this.getStyles();
+    const { isLoading, classObject, selectedRows, currentPage, items } = this.state;
+    const { dataEndpointName } = this.props.params;
 
     return (
       <div>
@@ -144,11 +140,10 @@ const DataEndpointsPreview = React.createClass({
             <DataObjectsTable
               withEditDialog={false}
               items={items}
-              hasNextPage={hasNextPage}
+              initialColumns={DataObjectsTableInitialColumns}
               selectedRows={selectedRows}
               classObject={classObject}
               handleRowSelection={this.handleRowSelection}
-              handleMoreRows={this.handleMoreRows}
             />
           </Loading>
         </Container>
